Support multiple rendering pass
The previous implementation of output() used a static float array for
the z-buffer, which caused rendering issues—such as broken animations:
when transitioning to a new scene due to residual values left in the
buffer. By switching to dynamically allocated z-buffers, we ensure that
the buffer is refreshed each time, eliminating artifacts from old data.